remove default config sync directory
Issue we are trying to fix is that we are seeing the changed config files from the prior import in the sync tab in the UI after we pull / checkout a branch with new config yml files.  Wondering if it is because the UI is defaulting to the wrong directory for config or if it is getting tripped up.